pyproject.toml
[tool.poetry]
name = "trade-csv-upload-api"
version = "0.1.0"
description = "Trade CSV Upload API with SQLModel"
authors = ["User <user@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
sqlmodel = "^0.0.8"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
asyncpg = "^0.26.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
pydantic = "^1.10.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
